Keep the provider outcome constant out of its type file
Same split as the other campaign constants: the value lives in a
.constant.ts and the type file only derives from it.
