+++ packages/twenty-server/src/engine/core-modules/emailing-domain/constants/campaign-provider-outcome.constant.ts
@@ -1,0 +1,8 @@
+export const CAMPAIGN_PROVIDER_OUTCOME = {
+  DELIVERED: 'DELIVERED',
+  SOFT_BOUNCED: 'SOFT_BOUNCED',
+  BOUNCED: 'BOUNCED',
+  COMPLAINED: 'COMPLAINED',
+  REJECTED: 'REJECTED',
+  RENDERING_FAILED: 'RENDERING_FAILED',
+} as const;
